Revert "Re-added the Contact.html"
This reverts commit 7c25c30ecc9ae09c57d30ec381ad1dac6755f832.

Revert the commit that readded the Contact.html file. If there were any
commits on this file id doesn't work to revert the commit that added it
but if there are no commits on this file or the commits were reverted it
will work to revert the commit that added the file.